vet: add custom go vet that detect clienterrors issues
The "WorkerClientError()" in osbuild-composer takes a
`details interface{}` parameter that is then directly marshaled
to json. if an `error' from e.g. `fmt.Errorf()` is passed this
will leads to silent dropping of the error. So detect this via
a custom `vet` module. Note that this will not inspect the type
recursively so it's a bit shallow.